+++ database/migrations/2024_08_11_084622_bookmark.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('bookmarks', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('internshipID');
-            $table->unsignedInteger('studentID');
+            $table->unsignedBigInteger('internshipID');
+            $table->unsignedBigInteger('studentID');
             $table->timestamps();
             $table->foreign('internshipID')->references('id')->on('internships');
             $table->foreign('studentID')->references('id')->on('students');
